refactor: reorganize AI-related modules and update imports
- Moved `scrapper` and `embeddings` modules to the `ai` directory for better structure.
- Updated import paths in `server.ts`, `server.test.ts`, and `bookmarks.ts` to reflect the new locations.
- Added new implementations for `embedText` and `getPageContent` in their respective files under the `ai` directory.
